add initial cpu fp8 kv cache impl, only support float32

+++ vllm/utils.py
@@ -335,7 +335,7 @@ def _generate_random_fp8(
     # Inf | N/A         | s.11111.00
     # NaN | s.1111.111  | s.11111.{01,10,11}
     from vllm import _custom_ops as ops
-    tensor_tmp = torch.empty_like(tensor, dtype=torch.float16)
+    tensor_tmp = torch.empty_like(tensor, dtype=torch.float32)
     tensor_tmp.uniform_(low, high)
     ops.convert_fp8(tensor, tensor_tmp)
     del tensor_tmp
